pyproject.toml
[tool.poetry]
name = "papurika"
version = "0.1.0"
description = ""
authors = ["Hanaasagi <ambiguous404@gmail.com>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.6"
grpcio = "^1.26.0"
grpcio-tools = "^1.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
